requirements.txt
pytorch-lightning==0.9.0
wandb==0.10.11
dotmap
pandas
nlpaug
matplotlib
torch_dct
torchaudio
